separating web server class
